\name{FEM-package}
\alias{FEM-package}
\alias{FEM}
\docType{package}
\title{
 FEM
}
\description{
Identifies interactome hotspots ofdifferential promoter methylation and differentialexpression, where an inverse association between methylation and gene expression is assumed
}
\details{
\tabular{ll}{
Package: \tab FEM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-22\cr
License: \tab GPL >2 \cr
}
}
\author{
Yinming Jiao and Andrew Teschendorff
Maintainer: Yinming Jiao  <20907099@zju.edu.cn>
}
\references{
1Jiao Y, Widschwendter M, Teschendorff AE. A systems-level integrative framework for genome-wide DNA methylation and gene expression data identifies differential gene expression modules under epigenetic control. Bioinformatics 2014, doi: 10.1093/bioinformatics/btu316 (2014-05-02)
2 Jones A, Teschendorff AE, Li Q, Hayward JD, Kannan A, et al. (2013) Role of dna methylation and epigenetic silencing of hand2 in endometrial cancer development. PLoS Med 10:e1001551.
3 Reichardt J, Bornholdt S (2006) Statistical mechanics of community detection. Phys Rev E 74:016110. doi:10.1103/PhysRevE.74.016110. URL http://link.aps.org/doi/10.1103/PhysRevE.74.016110.
4 West J, Beck S, Wang X, Teschendorff AE (2013) An integrative network algorithm identifies age-associated differential methylation interactome
hotspots targeting stem-cell differentiation pathways. Sci Rep 3:1630.


}


\keyword{ package }
\seealso{
}
\examples{

}
