% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciate.R
\name{spec_industry}
\alias{spec_industry}
\title{Speciation of emissions}
\usage{
spec_industry(x, spec = "GAS", code, usepa_code, list = FALSE)
}
\arguments{
\item{x}{Emissions estimation}

\item{spec}{Character any of "GAS" or "PM"}

\item{code}{Character any of "01_petroleum", "02_portland_cement", "03_paper_cellulose",
"04_bagasse", "05_natural_gas", "06_coal" OR "07_fuel_oil"}

\item{usepa_code}{Character any of "4713", "3011", "8799", "SUG02", "2442", "1185", "0001",
"13505C", "91127", "4733", "SUGP02", "1120310" OR "91114"}

\item{list}{when TRUE returns a list with number of elements of the list as the number species of pollutants}
}
\value{
dataframe or list
}
\description{
\code{speciate} separates emissions in different compounds.
It covers black carbon and organic matter from particulate matter. Soon it
will be added more speciations
}
\examples{
{
# Do not run
pm <- rnorm(n = 100, mean = 400, sd = 2)
(df <- spec_industry(pm, code = "01_petroleum"))
}
}
\references{
Simon, H., Beck, L., Bhave, P. V., Divita, F., Hsu, Y., Luecken, D., ...
Strum, M. (2010). The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), 196-206.
}
