\name{FEM-package}
\alias{FEM-package}
\alias{FEM}
\docType{package}
\title{
Identification of interactome hotspots of differential promoter DNA methylation
and differential expression
}
\description{
The FEM package performs a systems-level integrative analysis of DNA
methylationa and gene expression. It seeks modules of functionally
related genes which exhibit differential promoter DNA methylation and
differential expression, where an inverse association between promoter
DNA methylation and gene expression is assumed. For full details, see
Jiao et al Bioinformatics 2014.
}
\details{
\tabular{ll}{
Package: \tab FEM\cr
Type: \tab Package\cr
Version: \tab 2.1.4\cr
Date: \tab 2015-03-27\cr
License: \tab GPL >2 \cr
}
}
\author{
"Yinming Jiao"<20907099@zju.edu.cn>, "Andrew E Teschendorff"<andrew@picb.ac.cn>
}
\references{
1) Jiao Y, Widschwendter M, Teschendorff AE. A systems-level integrative
framework for genome-wide DNA methylation and gene expression data
identifies differential gene expression modules under epigenetic
control. Bioinformatics 2014 Aug 15;30(16):2360-2366.

2) Jones A, Teschendorff AE, Li Q, Hayward JD, Kannan A, et al. (2013) Role of dna methylation and epigenetic silencing of hand2 in endometrial cancer development. PLoS Med 10:e1001551.

3) Reichardt J, Bornholdt S (2006) Statistical mechanics of community detection. Phys Rev E 74:016110. doi:10.1103/PhysRevE.74.016110. URL http://link.aps.org/doi/10.1103/PhysRevE.74.016110.

4) West J, Beck S, Wang X, Teschendorff AE (2013) An integrative network algorithm identifies age-associated differential methylation interactome
hotspots targeting stem-cell differentiation pathways. Sci Rep 3:1630.


}


\keyword{ package }
\seealso{
}
\examples{

}
